+++ rpgram/characters/__init__.py
@@ -1,2 +1,3 @@
 from rpgram.characters.char_base import BaseCharacter
 from rpgram.characters.char_player import PlayerCharacter
+from rpgram.characters.char_non_player import NPCharacter
